requirements.txt
streamlit
opencv-python
tensorflow
keras
numpy
Pillow
requests
